feat: メニュー更新処理をサーバー側に移動
